requirements.txt
PyCryptodome~=3.15.0
